+++ MatrixSDK/Data/MXRoomSummaryUpdater.m
@@ -2,6 +2,7 @@
  Copyright 2017 OpenMarket Ltd
  Copyright 2017 Vector Creations Ltd
  Copyright 2018 New Vector Ltd
+ Copyright 2019 The Matrix.org Foundation C.I.C
 
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
@@ -58,6 +59,11 @@
                 NSLog(@"[MXRoomSummaryUpdater] updateRoomSummary: Cannot reset last message after redaction. Room: %@", summary.roomId);
             } commit:YES];
         }
+        return NO;
+    }
+    else if (event.isEditEvent)
+    {
+        // Do not display update events in the summary
         return NO;
     }